Got things working
MainWebViewPanel.ts: added "data" in the csp to allow external fonts and images.

Added index.css to avoid warnings.

Added react-router. This is instead of react-router-dom which is depreciated.

Added some of the CreateWithImagination components and page

main.tsx: do not know what the default homepage is, so if it isn't listed in the routes, reroute to "/".

Re-did navigation throughout old code to work with react-router.

Added other files/code from old code.

--

TODO: fix some CSS and add the Sketch components.